Conflicts:
	target/classes/META-INF/maven/com.saidashevar.ptgame/pixel-tactics-game/pom.properties

+++ src/main/java/com/saidashevar/ptgame/repository/CardRepository.java
@@ -2,6 +2,6 @@ package com.saidashevar.ptgame.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.saidashevar.ptgame.model.Card;
+import com.saidashevar.ptgame.model.cards.Card;
 
-public interface CardRepository extends JpaRepository<Card, Long> {}
+public interface CardRepository extends JpaRepository<Card, Integer> {}
